// src/interfaces/FightersApiInterfaces.tsx
// Generated by https://quicktype.io

export interface Universe {
  objectID: string;
  name: string;
  description: string;
}

// Generated by https://quicktype.io

export interface Fighter {
  objectID: string;
  name: string;
  universe: string;
  price: string;
  popular: boolean;
  rate: number;
  downloads: string;
  description: string;
  created_at: string;
  imageURL: string;
}